can now re-run for single parameters, support keyword argument based caching

+++ src/slurm_map.py
@@ -2,38 +2,73 @@ import json
 import os, sys
 import subprocess
 import time
+import inspect
+import hashlib
 from typing import Callable, List, Any
+from pathlib import Path
 
 import dill as pickle
+import rich
 
 from utils import robust_rmtree, unpickleWithTimeout, jobs_running, watchFileAsync, namedTemporaryFile
 
 SLURM_MAP_DIR = ".slurm_map"
 
+config = {"print": "rich"}
+
+def print(*args, **kwargs):
+    if config["print"] == "rich":
+        rich.print("[grey50](slurm-map):", *args, **kwargs)
+    elif config["print"] == "plain":
+        print(*args, **kwargs)
+    else:
+        pass
+
 stopListeners = False
 
-def startJobs(folder: str, function: Callable, data: List[Any], slurm_args: str, extra_commands: List[str]) -> None:
+def startJobs(folder: str, function: Callable, data: List[Any], slurm_args: str, extra_commands: List[str], kwargs) -> None:
     with namedTemporaryFile(folder, "function.dill") as function_file:
         pickle.dump(function, function_file, recurse=True)
 
+    with namedTemporaryFile(folder, "kwargs.json", "w") as global_args_file:
+        json.dump(kwargs, global_args_file)
+
     job_ids = []
 
-    for i, x in enumerate(data):
-        with namedTemporaryFile(folder, f"arg_{i}.dill") as arg_file:
+
+    Path(os.path.join(folder, "slurm")).mkdir(parents=True, exist_ok=True)
+    Path(os.path.join(folder, "results")).mkdir(parents=True, exist_ok=True)
+
+    # Workaround: periodically call ls in order to force any underlying filesystem to update. 
+    subprocess.Popen(['ls', os.path.join(folder, "results")], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+    subprocess.Popen(['ls', os.path.join(folder, "args")], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+
+    len_new_data = 0
+    for x in data:
+        data_hash = hashlib.md5(pickle.dumps(x)).hexdigest()
+
+        results_exist = os.path.isfile(os.path.join(folder, "results", f"res_{data_hash}.dill"))
+        args_exist = os.path.isfile(os.path.join(folder, "args", f"arg_{data_hash}.dill"))
+        if args_exist or results_exist: #Already have results for this data element
+            continue 
+    
+        len_new_data += 1
+
+        with namedTemporaryFile(os.path.join(folder, "args"), f"arg_{data_hash}.dill") as arg_file:
             pickle.dump(x, arg_file)
 
-        with namedTemporaryFile(folder, f"run_{i}.slurm", mode="w") as slurm_file:
+        with namedTemporaryFile(folder, f"run_{data_hash}.slurm", mode="w") as slurm_file:
             slurm_file_contents = [
                 "#!/bin/sh",
-                f"#SBATCH --output " + os.path.join(folder, f"job_{i}.log"),
-                f"#SBATCH --error " + os.path.join(folder, f"job_{i}.log")
+                f"#SBATCH --output " + os.path.join(folder, "slurm", f"job_%j.log"),
+                f"#SBATCH --error " + os.path.join(folder, "slurm", f"job_%j.log")
             ]
             lines = slurm_file_contents + extra_commands
             slurm_file.writelines([command + "\n" for command in lines])
 
             python_executable = sys.executable
             # Start *this* file as a slurm job, passing it the pickled function, args and results file
-            python_command = f"{python_executable} -u {os.path.abspath(__file__)} execute {function_file.name} {arg_file.name} {os.path.join(folder, f'res_{i}.dill')}\n"
+            python_command = f"{python_executable} -u {os.path.abspath(__file__)} execute {function_file.name} {arg_file.name} {global_args_file.name} {os.path.join(folder, 'results', f'res_{data_hash}.dill')}\n"
             slurm_file.writelines([python_command])
 
         
@@ -46,20 +81,35 @@ def startJobs(folder: str, function: Callable, data: List[Any], slurm_args: str,
             else:
                 print(line_str)
 
-    if len(job_ids) == len(data):
+    if len_new_data < len(data):
+        print(f"Re-using {len(data) - len_new_data} results -> only need {len_new_data} new jobs")
+    if len(job_ids) == len_new_data:
         print(f"Submitted {len(job_ids)} jobs.")
+    else:
+        print(f"Warning! There was a problem with job submission. Only submitted {len(job_ids)} out of {len_new_data} jobs!")    
 
+    previous_job_ids = []
+    try:
+        with namedTemporaryFile(folder, "job_ids.json", mode="r") as f:
+            previous_job_ids = json.load(f)
+    except Exception:
+        pass
     with namedTemporaryFile(folder, "job_ids.json", mode="w") as f:
-        json.dump(job_ids, f)
+        json.dump(previous_job_ids + job_ids, f)
 
 
-def map(function: Callable, data: List[Any], slurm_args: str = None, extra_commands: List[str] = None, cleanup=True, folder=None) -> List[Any]:
+def map(function: Callable, data: List[Any], slurm_args: str = None, extra_commands: List[str] = None, cleanup=True, kwargs=None) -> List[Any]:
     """ 
         Maps function over data, where every element of data is executed as a separate slurm batch job.
+
+        The result is equivalent to 
+
+        [function(x, **kwargs) for x in data]
         
         slurm_args: additional arguments to pass to slurm when submitting the job
         extra_commands: additional commands to run in the job before starting the evaluation of function. (Eg. 'module load python3')
         cleanup: If true, delete temporary files created after this function finishes. (Temporary files are pickled versions of the function, data and results, as well as the slurm bash scripts and log files.) 
+        args: Additional keyword arguments for 'function'
 
     """
 
@@ -68,23 +118,21 @@ def map(function: Callable, data: List[Any], slurm_args: str = None, extra_comma
     if extra_commands is None:
         extra_commands = []
 
-    folder_arg = folder
+    callee_filename = os.path.basename(inspect.stack()[1].filename).rstrip(".py")
 
-    if folder is None:
-        folder = os.path.join(".", SLURM_MAP_DIR, function.__name__) #f"./{SLURM_MAP_DIR}/{function.__name__}/")
-    else:
-        folder = os.path.join(".", SLURM_MAP_DIR, folder, function.__name__)
+    args_str = json.dumps(kwargs)
+    args_hash = hashlib.md5(args_str.encode('utf-8')).hexdigest()
 
-    if os.path.isfile(os.path.join(folder, "job_ids.json")):
-        print("Jobs were already started previously. Reusing those results.")
-    else: startJobs(folder, function, data, slurm_args, extra_commands)
+    folder = os.path.join(".", SLURM_MAP_DIR, callee_filename, function.__name__, args_hash)
+
+    startJobs(folder, function, data, slurm_args, extra_commands, kwargs)
 
     with open(os.path.join(folder, "job_ids.json")) as f:
         job_ids = json.load(f)
     
-    print("Waiting for results from jobs...")
-    
     running = jobs_running(job_ids)
+
+    print(f"Waiting for {sum(running)} jobs to finish...")
 
     global stopListeners
     stopListeners = False
@@ -93,7 +141,7 @@ def map(function: Callable, data: List[Any], slurm_args: str = None, extra_comma
         return stopListeners
 
     # Start file watchers for unfinished jobs in order to get stdout.
-    log_watchers = [watchFileAsync(os.path.join(folder, f"job_{i}.log"), stopListenerCallback) if r else None for i, r in enumerate(running)]
+    log_watchers = [watchFileAsync(os.path.join(folder, "slurm", f"job_{id}.log"), stopListenerCallback) if r else None for id, r in zip(job_ids, running)]
 
     # Wait until all jobs are finished
     try: 
@@ -102,39 +150,36 @@ def map(function: Callable, data: List[Any], slurm_args: str = None, extra_comma
             running = jobs_running(job_ids)
     except KeyboardInterrupt:
         stopListeners = True
-        temp = function.__name__
-        if folder_arg is not None:
-            temp = os.path.join(folder_arg, function.__name__)
-        print(f"\nStopping to wait for results. Jobs will continue to run. Use 'python -m slurm_map cancel {temp}' to cancel the jobs.")
+        print(f"\nStopping to wait for results. Jobs will continue to run. Use 'python -m slurm_map cancel {folder}' to cancel the jobs.")
         exit()
     
-    print("Jobs done, waiting for results...", end="")
+    print("Jobs done, waiting for results...")
 
     # collect results
     results = [None]*len(data)
     succeeded = [False]*len(data)
     iteration = 0
-    max_iterations = 10
+    max_iterations = 20
 
     while (not all(succeeded)) and (iteration < max_iterations):
         for i in [j for j in range(len(data)) if not succeeded[j]]:
             try:
-                with open(os.path.join(folder, f"res_{i}.dill"), "rb") as f:
+                data_hash = hashlib.md5(pickle.dumps(data[i])).hexdigest()
+                with open(os.path.join(folder, "results", f"res_{data_hash}.dill"), "rb") as f:
                     results[i] = pickle.load(f)
                 succeeded[i] = True
-            except Exception:
+            except Exception as e:
                 pass
+        # Workaround: periodically call ls in order to force any underlying filesystem to update. 
+        subprocess.Popen(['ls', os.path.join(folder, "results")], stdout=subprocess.PIPE, stderr=subprocess.PIPE)
         time.sleep(0.5)
         iteration += 1
 
     all_succeeded = all(succeeded)
     if all_succeeded:
-        print(" Done.")
+        print("Done.")
     else:
-        temp = function.__name__
-        if folder_arg is not None:
-            temp = os.path.join(folder_arg, function.__name__)
-        print(f"\nWarning: not all jobs returned a result. Not cleaning up. Clean up manually using 'python -m slurm_map cleanup {temp}'")
+        print(f"Warning: not all jobs returned a result. Not cleaning up. Clean up manually using 'python -m slurm_map cleanup {folder}'")
 
     if cleanup and all_succeeded:
         robust_rmtree(folder)
@@ -142,19 +187,26 @@ def map(function: Callable, data: List[Any], slurm_args: str = None, extra_comma
     return results
 
 
-def compute(function_file: str, arg_file: str, res_file: str) -> None:
+def compute(function_file: str, arg_file: str, kwarg_file:str, res_file: str) -> None:
     function = unpickleWithTimeout(function_file)
     args = unpickleWithTimeout(arg_file)
 
+    with open(kwarg_file) as f:
+        kwargs = json.load(f)
+
+    if kwargs is None:
+        kwargs = {}
+
     try:
-        os.remove(arg_file)
-    except FileNotFoundError:
-        pass
+        res = function(args, **kwargs)
 
-    res = function(args)
-
-    with open(res_file, "wb") as f:
-        pickle.dump(res, f)
+        with open(res_file, "wb") as f:
+            pickle.dump(res, f)
+    finally:
+        try: # Remove arg_file even when an error occurs. That way, a second slum_map call will retry this call.
+            os.remove(arg_file)
+        except FileNotFoundError:
+            pass
 
     print("slurm_map saved data!")
 
@@ -185,20 +237,22 @@ if __name__ == "__main__":
     parser_execute.add_argument(
         'arg_file', help='Path of the .dill file encoding the arguments of the function to be executed.')
     parser_execute.add_argument(
+        'kwarg_file', help='Path of the .dill file encoding the keyword arguments of the function to be executed.')
+    parser_execute.add_argument(
         'res_file', help='Path of the .dill file where the results of the function shall be stored.')
 
     args = parser.parse_args()
     if args.action == 'execute':
-        compute(args.function_file, args.arg_file, args.res_file)
+        compute(args.function_file, args.arg_file, args.kwarg_file, args.res_file)
     elif args.action == 'cleanup':
-        path = os.path.join(".", SLURM_MAP_DIR, args.function_name)
+        path = os.path.join(args.function_name)
         if os.path.isdir(path):
             robust_rmtree(path)
             print("Cleaned up", path)
         else:
             print(f"Nothing to do. {path} does not exist.")
     elif args.action == 'cancel':
-        path = os.path.join(".", SLURM_MAP_DIR, args.function_name, "job_ids.json")
+        path = os.path.join(args.function_name, "job_ids.json")
         if os.path.isfile(path):
             with open(path) as f:
                 job_ids = json.load(f)
